Added Database.watch() which maintains a local cache and makes callbacks on changes
This will improve on stream() by avoiding triggering the callback
every hour when we reconnect following expiration of the ID token.
Right now it fires childAdded when that happens, but the cache
will allow fixing that.

As a side-effect, it also allows instant local access to the data,
though there isn't any nice API for that yet.